refactor(aws-lambda): refactor aws-lambda plugin code with lua-resty-aws (#11350)
* refactor(aws-lambda): add lua-resty-aws library and libexpat dependencies into rockspec and bazel configs

The commit introduces latest version of lua-resty-aws as Kong's
dependency. Since lua-resty-aws relies on luaexpat to do xml decoding,
Kong also needs to build libexpat during compiling/packaging.

* refactor(aws-lambda): use lua-resty-aws and rewrite fetch credential

The commits rewrite part of the aws-lambda plugin code so that the IAM
role credential fetching is replaced by using lua-resty-aws credential
provider.

* refactor(aws-lambda): refactor aws-lambda plugin

This commits does refactoring on the majority of the aws-lambda plugin
code. The IAM role credential fetching and lambda function invoking has
been replaced by using the lua-resty-aws library directly.

* style(*): remove useless lua file

* fix(cd): fix explain manifest for libexpat

* fix(cd): fix buildifier style

* fix(*): try to fix lambda plugin init_worker

* fix(*): fix http proxy & sts regional endpoint config

* fix(*): execute plugin init code correctly

* fix(*): remove lambda returned content length

* chore(*): move libexpat from cross_deps to standalone repo

* fix(*): do not override global config credential

* chore(*): remove non-debug flag

* chore(*): bump lua-resty-aws version to 1.3.0
