+++ .requirements
@@ -4,6 +4,7 @@ OPENRESTY=1.21.4.1
 LUAROCKS=3.9.2
 OPENSSL=3.1.4
 PCRE=8.45
+LIBEXPAT=2.5.0
 
 LUA_KONG_NGINX_MODULE=4d19e8d19c6dbc07eba5cf6f5ebacad95266f928 # 0.6.0
 LUA_RESTY_LMDB=951926f20b674a0622236a0e331b359df1c02d9b # 1.3.0
