+++ changelog/unreleased/kong/lua-resty-aws-bump.yml
@@ -1,0 +1,2 @@
+message: "Bumped lua-resty-aws from 1.2.3 to 1.3.0"
+type: dependency
