+++ kong/init.lua
@@ -628,6 +628,8 @@ function Kong.init()
   -- Load plugins as late as possible so that everything is set up
   assert(db.plugins:load_plugin_schemas(config.loaded_plugins))
 
+  assert(db.plugins:execute_plugin_init())
+
   if is_stream_module then
     stream_api.load_handlers()
   end
