+++ scripts/explain_manifest/fixtures/el7-amd64.txt
@@ -55,6 +55,12 @@
   - libc.so.6
   Runpath   : /usr/local/kong/lib
 
+- Path      : /usr/local/kong/lib/libexpat.so.1.8.10
+  Needed    :
+  - libstdc++.so.6
+  - libm.so.6
+  - libc.so.6
+
 - Path      : /usr/local/kong/lib/libssl.so.3
   Needed    :
   - libstdc++.so.6
@@ -95,6 +101,12 @@
 
 - Path      : /usr/local/lib/lua/5.1/lua_system_constants.so
   Needed    :
+  - libc.so.6
+  Rpath     : /usr/local/kong/lib
+
+- Path      : /usr/local/lib/lua/5.1/lxp.so
+  Needed    :
+  - libexpat.so.1
   - libc.so.6
   Rpath     : /usr/local/kong/lib
